+++ internal/tracing/setup/config.go
@@ -61,7 +61,7 @@ func SetFileWriter(name string) error {
 
 // SetEndpoint supplies the endpoint to the trace sink service for the
 // 'production' trace exporter variant
-func SetEndpoint(host string, port int) error {
+func SetEndpoint(host string, port uint16) error {
 	endpoint := fmt.Sprintf("%s:%d", host, port)
 	return production.SetEndpoint(endpoint, grpc.WithInsecure())
 }
